requirements.txt
beautifulsoup4==4.13.4
httpx==0.28.1
aiofiles==23.2.1
pandas==2.2.3

# NotebookLM MCP integration
mcp>=1.0.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
